% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{add_ss_indices}
\alias{add_ss_indices}
\title{Add incidences corresponding to state space}
\usage{
add_ss_indices(dat, ss)
}
\arguments{
\item{dat}{a data.frame with particular columns coded in particular ways}

\item{ss}{state space definition (fp$ss)}
}
\description{
Add incidences corresponding to state space
}
\examples{
\dontrun{
data(survey_hts)
dat <- subset(survey_hts, survey_hts$country == "Malawi" & outcome == "evertest")
df <- add_ss_indices(dat, fp$ss)
}
}
