% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-pjnz.R
\name{calc_asfr}
\alias{calc_asfr}
\title{Calculate ASFR from TFR and fertility distribution}
\usage{
calc_asfr(tfr, asfd)
}
\arguments{
\item{tfr}{vector of annual TFR values}

\item{asfd}{array of proportion of births by 5 year age group 15-49}
}
\value{
array of age-specific fertility rate by single-year of age 15-49.
}
\description{
Calculate ASFR from TFR and fertility distribution
}
