% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-pjnz.R
\name{combine_inputs}
\alias{combine_inputs}
\title{Combine PJNZ inputs}
\usage{
combine_inputs(lst)
}
\arguments{
\item{lst}{a list of inputs, each returned from extract_pjnz()}
}
\description{
Combine PJNZ inputs
}
