% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fp.R
\name{create_beers}
\alias{create_beers}
\title{Beers coefficients to distribute from 5-year to single-year of age}
\usage{
create_beers(n5yr)
}
\description{
Beers coefficients to distribute from 5-year to single-year of age
}
