% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hts_param.R
\name{create_hts_param}
\alias{create_hts_param}
\title{Create HIV testing rates from parameter inputs}
\usage{
create_hts_param(theta, fp)
}
\description{
Create HIV testing rates from parameter inputs
}
\details{
This function creates two arrays \code{hts_rate} and \code{diagn_rate}.
The \code{hts_rate} array summarizes HIV testing rates for the HIV negative population stratified by:
\itemize{
\item HIV aggregated age groups (hAG: 1-9)
\item sex (1 = male; 2 = female)
\item testing history (1 = never tested; 2 = previously tested negative)
\item year
}

Array \code{diagn_rate} summarizes HIV testing rates for the HIV positive population stratified by:
\itemize{
\item CD4 stage (hDS: 1-7)
\item HIV age group (hAG: 1-9)
\item sex (1 = male; 2 = female)
\item testing and treatment history (1 = never tested; 2 = previously tested negative; 3 = diagnosed, not treated; 4 = on ART)
\item year
}
}
