% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{evertest}
\alias{evertest}
\title{Proportion ever tested by age, sex, or HIV status}
\usage{
evertest(mod, fp, df, VERSION = "C")
}
\arguments{
\item{mod}{simulation model output}

\item{fp}{simulation model parameter inputs}

\item{df}{a data.frame with indices for prediction, see Details.}
}
\value{
a vector
}
\description{
This function calculates proportion of ever tested among a population
stratified by age group, sex, HIV status, and year.
}
\details{
Age groups are specified in terms of aggregate HIV age groups: 15-16, 17-19, 20-24, ..., 45-49, 50+.
Another function could be added to handle other age groups if needed, with additional computational
complexity.

The argument df must contain the following columns:
\itemize{
\item haidx:   HIV age group (1 = 15-16, 2 = 17-19, 3 = 20-24, ..., 8 = 45-49, 9 = 50+)
\item sidx:    sex (1 = male, 2 = female, 0 = both)
\item hvidx:   HIV status (1 = negative, 2 = positive, 0 = all)
\item yidx:    year index
\item hagspan: number of HIV age groups to span
}
}
\examples{

\dontrun{
data(survey_hts)
dat <- subset(survey_hts, country == "Malawi" & outcome == "evertest")
df <- add_ss_indices(dat, fp$ss)

df$pred <- evertest(mod, fp, df)
}

}
