% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-pjnz.R
\name{extract_pjnz}
\alias{extract_pjnz}
\title{Extract outputs from PJNZ needed for first90 model}
\usage{
extract_pjnz(pjnz = NULL, dp_file = NULL, pjn_file = NULL)
}
\arguments{
\item{pjnz}{filepath to PJNZ file}

\item{dp_file}{filepath to a .DP file}

\item{pjn_file}{filepath to a .PJN file}
}
\value{
a list
}
\description{
Extract outputs from PJNZ needed for first90 model
}
