% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-pjnz.R
\name{get_dp_Sx}
\alias{get_dp_Sx}
\title{Get mortality probability (Sx) from Spectrum DP file}
\usage{
get_dp_Sx(dp, proj_years)
}
\description{
Get mortality probability (Sx) from Spectrum DP file
}
\details{
This function extracts Sx for ages 0:79 and 80+. Spectrum calculates a
separate Sx for age 80. The population projection model in EPP-ASM needs
to be updated to handle this.
}
