% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-pjnz.R
\name{get_pjn_country}
\alias{get_pjn_country}
\title{Get country name from parsed PJN}
\usage{
get_pjn_country(pjn)
}
\arguments{
\item{pjn}{parsed PJN file}
}
\description{
Get country name from parsed PJN
}
\details{
\code{pjn} should be via \code{read.csv(pjn_file, as.is = TRUE)}
}
