% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-pjnz.R
\name{get_pjn_region}
\alias{get_pjn_region}
\title{Get subnational region from parsed PJN}
\usage{
get_pjn_region(pjn)
}
\arguments{
\item{pjn}{parsed PJN file}
}
\description{
Get subnational region from parsed PJN
}
\details{
\code{pjn} should be via \code{first90_read_csv_character(pjn_file)}
}
