% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{number_diagnoses}
\alias{number_diagnoses}
\title{Number of new diagnoses by age and sex}
\usage{
number_diagnoses(mod, fp, df, VERSION = "R")
}
\value{
A data frame reporting new diagnoses in two ways:
\enumerate{
\item Approximated from the testing rates and mid-year populations
as in the function \code{\link[=number_tests]{number_tests()}}.
\item Based on the number of new diagnoses recorded in each model time step.
}

The column \code{late_diagnoses} reports the number who are initiated to ART
directly from the undiagnosed population.

The two approaches are largely for debugging purposes to understand
the implications of the model choices.
}
\description{
Number of new diagnoses by age and sex
}
