% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retest.R
\name{number_retests}
\alias{number_retests}
\title{Calculate number adn proportion of retests conducted or number tested by year
funtion based on number test in the outputs.R script}
\usage{
number_retests(mod, fp, df)
}
\arguments{
\item{mod}{model output of class 'eppasm'}

\item{fp}{parameter inputs (class 'specfp')}

\item{df}{a data.frame with indices for prediction. See \code{\link[=evertest]{evertest()}} for more information.}
}
\value{
a data.frame consisting of the number of tests, and population size corresponding to rows of df
}
\description{
Calculate number adn proportion of retests conducted or number tested by year
funtion based on number test in the outputs.R script
}
\details{
Number of tests (or number tested in past 12 months) are approximated by mid-year
counts and annual testing rates within each stratum.
}
