% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{number_tests}
\alias{number_tests}
\title{Calculate number of tests conducted or number tested by year}
\usage{
number_tests(mod, fp, df, tested12m = FALSE, VERSION = "C")
}
\arguments{
\item{mod}{model output of class 'eppasm'}

\item{fp}{parameter inputs (class 'specfp')}

\item{df}{a data.frame with indices for prediction. See \code{\link[=evertest]{evertest()}} for more information.}
}
\value{
a data.frame consisting of the number of tests, number tested in the past 12 months, and population size corresponding to rows of df
}
\description{
Calculate number of tests conducted or number tested by year
}
\details{
Number of tests (or number tested in past 12 months) are approximated by mid-year
counts and annual testing rates within each stratum.
}
