% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{prepare_inputs_from_extracts}
\alias{prepare_inputs_from_extracts}
\title{Create model inputs from aggregated Spectrum PJNZ}
\usage{
prepare_inputs_from_extracts(pjnz_in)
}
\arguments{
\item{pjnz_in}{a list of outputs from extract_pjnz}
}
\description{
Create model inputs from aggregated Spectrum PJNZ
}
