% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{svy_hts_interface}
\alias{svy_hts_interface}
\title{Function to present the survey data on shinny interface}
\usage{
svy_hts_interface(survey_hts, cnt, age_group = c("15-24", "25-34", "35-49"))
}
\description{
Function to present the survey data on shinny interface
}
