% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{select_prgmdata}
\alias{select_prgmdata}
\title{Process programmatic data on number of tests}
\usage{
select_prgmdata(prgm_dat, cnt, age_group, prgm_dat_sex = NULL)
}
\description{
Process programmatic data on number of tests
}
