% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_output.R
\name{spectrum_output_table}
\alias{spectrum_output_table}
\title{Table to write CSV outputs for Spectrum}
\usage{
spectrum_output_table(mod, fp)
}
\value{
a data.frame to write to CSV file for ingestion into Spectrum
}
\description{
Table to write CSV outputs for Spectrum
}
\details{
Presently this returns point estimates for age 15+ population by sex:
\itemize{
\item Number PLHIV,
\item Ever tested among PLHIV
\item Aware of HIV+ status
\item On ART
}
}
\examples{

write.csv(spectrum_output_table(mod, fp), 
          "~/Downloads/Malawi-shiny90-example-output.csv", 
           row.names = FALSE)
}
